+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Default to smallest instance type for asg
+* Move to graviton for asg
+
 # 3.9.0
 
 * updating aurora version to 5.7.mysql_aurora.2.11.1